Make offset in loop device helper uint64_t
It is actually never used other than for 0, but
it should be correct.

+++ lib/utils_loop.h
@@ -9,11 +9,13 @@
 #ifndef _UTILS_LOOP_H
 #define _UTILS_LOOP_H
 
+#include <stdint.h>
+
 /* loopback device helpers */
 
 char *crypt_loop_backing_file(const char *loop);
 int crypt_loop_device(const char *loop);
-int crypt_loop_attach(char **loop, const char *file, int offset,
+int crypt_loop_attach(char **loop, const char *file, uint64_t offset,
 		      int autoclear, int *readonly, size_t blocksize);
 int crypt_loop_detach(const char *loop);
 int crypt_loop_resize(const char *loop);
